telegram_bot/db.py
"""
Módulo de conexión a la base de datos PostgreSQL.
Provee un context manager para obtener cursores con auto-commit/rollback.
"""

import os
import psycopg2
from contextlib import contextmanager
from dotenv import load_dotenv

load_dotenv(os.path.join(os.path.dirname(__file__), ".env"))

DB_CONFIG = {
    "host": os.getenv("DB_HOST"),
    "database": os.getenv("DB_NAME"),
    "user": os.getenv("DB_USER"),
    "password": os.getenv("DB_PASSWORD"),
    "port": int(os.getenv("DB_PORT", 5432)),
}


@contextmanager
def get_cursor():
    """Context manager que abre conexión, ejecuta, y hace commit/rollback."""
    conn = psycopg2.connect(**DB_CONFIG)
    try:
        with conn.cursor() as cur:
            yield cur
        conn.commit()
    except Exception:
        conn.rollback()
        raise
    finally:
        conn.close()
