telegram_bot/requirements.txt
python-telegram-bot==21.10
psycopg2-binary==2.9.11
python-dotenv==1.1.1
